% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hues-package.R
\docType{package}
\name{hues}
\alias{hues}
\alias{hues-package}
\title{hues: Generate iwanthue Colour Palettes in R}
\description{
Based on \href{http://tools.medialab.sciences-po.fr/iwanthue/}{iwanthue},
\code{hues} provides functions that generate and plot palettes of distinct
colours through k-means clustering of LAB colour space.
}

