% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hues-package.R
\docType{package}
\name{hues}
\alias{hues}
\title{hues: Generate optimally distinct colour palettes}
\description{
This package generates and plot palettes of optimally distinct colours
in perceptually uniform colour space, based on
\href{http://tools.medialab.sciences-po.fr/iwanthue/}{iwanthue}. This is
done through k-means clustering of CIE Lab colour space, according to
user-selected constraints on hue, chroma, and lightness.
}
\references{
\itemize{
\item \href{http://tools.medialab.sciences-po.fr/iwanthue/}{iwanthue}
\item \href{https://github.com/medialab/iwanthue}{iwanthue GitHub repository}
}
}
